+++ examples/namespace.js
@@ -4,5 +4,3 @@ org.myapp.util = (function () {
   var util = {};
   return util;
 }());
-
-console.log(org);