+++ tests/browser/scripts/modules/b.js
@@ -1,5 +1,9 @@
 'use strict';
 
 $module(function (c, d, exports) { 
+
+  $namespace('shared');
+
   exports.done = true;
+
 }, ['c', 'd']);