+++ tests/node/modules/a.js
@@ -1,5 +1,8 @@
 'use strict';
 
 $module(function (c, exports) { 
+  
+  $namespace('shared');
+
   exports.done = true;
 }, ['c']);