+++ tests/node/modules/d.js
@@ -1,5 +1,8 @@
 'use strict';
 
 $module(function (exports) { 
+
+  $namespace('shared');
+  
   exports.done = true;
 });