+++ tests/node/modules/main.js
@@ -23,5 +23,9 @@ var test = { done: true };
 
 $module(function (url, a, b, t) { 
   console.log('in main, url=%j, a=%j, b=%j, t=%j', typeof url !== "undefined", a.done, b.done, t.done);
+
+  console.assert(typeof a.shared !== 'undefined');
+  console.assert(typeof b.shared !== 'undefined');
+
 }, ['url', 'a', 'b', test]);